requirements.txt
python-telegram-bot
fastapi
uvicorn[standard]
python-dotenv
requests